MF-573 - Restructure things api
