+++ things/api/http/groups/endpoint_test.go
@@ -1,0 +1,196 @@
+package groups_test
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"io"
+	"net/http"
+	"net/http/httptest"
+	"strconv"
+	"strings"
+	"testing"
+
+	"github.com/MainfluxLabs/mainflux/auth"
+	"github.com/MainfluxLabs/mainflux/logger"
+	"github.com/MainfluxLabs/mainflux/pkg/apiutil"
+	"github.com/MainfluxLabs/mainflux/pkg/mocks"
+	"github.com/MainfluxLabs/mainflux/pkg/uuid"
+	"github.com/MainfluxLabs/mainflux/things"
+	httpapi "github.com/MainfluxLabs/mainflux/things/api/http"
+	thmocks "github.com/MainfluxLabs/mainflux/things/mocks"
+	"github.com/MainfluxLabs/mainflux/users"
+	"github.com/opentracing/opentracing-go/mocktracer"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+const (
+	contentType    = "application/json"
+	email          = "user@example.com"
+	adminEmail     = "admin@example.com"
+	otherUserEmail = "other_user@example.com"
+	token          = email
+	wrongValue     = "wrong_value"
+	emptyValue     = ""
+	password       = "password"
+	orgID          = "374106f7-030e-4881-8ab0-151195c29f92"
+	orgID2         = "374106f7-030e-4881-8ab0-151195c29f93"
+)
+
+var (
+	user      = users.User{ID: "574106f7-030e-4881-8ab0-151195c29f94", Email: email, Password: password, Role: auth.Owner}
+	otherUser = users.User{ID: "ecf9e48b-ba3b-41c4-82a9-72e063b17868", Email: otherUserEmail, Password: password, Role: auth.Editor}
+	admin     = users.User{ID: "2e248e36-2d26-46ea-97b0-1e38d674cbe4", Email: adminEmail, Password: password, Role: auth.RootSub}
+	usersList = []users.User{admin, user, otherUser}
+	orgsList  = []auth.Org{{ID: orgID, OwnerID: user.ID}, {ID: orgID2, OwnerID: user.ID}}
+)
+
+type testRequest struct {
+	client      *http.Client
+	method      string
+	url         string
+	contentType string
+	key         string
+	token       string
+	body        io.Reader
+}
+
+func (tr testRequest) make() (*http.Response, error) {
+	req, err := http.NewRequest(tr.method, tr.url, tr.body)
+	if err != nil {
+		return nil, err
+	}
+	if tr.key != "" {
+		req.Header.Set("Authorization", apiutil.ThingPrefix+tr.key)
+	}
+	if tr.token != "" {
+		req.Header.Set("Authorization", apiutil.BearerPrefix+tr.token)
+	}
+	if tr.contentType != "" {
+		req.Header.Set("Content-Type", tr.contentType)
+	}
+	return tr.client.Do(req)
+}
+
+func newService() things.Service {
+	auth := mocks.NewAuthService(admin.ID, usersList, orgsList)
+	thingsRepo := thmocks.NewThingRepository()
+	profilesRepo := thmocks.NewProfileRepository(thingsRepo)
+	groupMembersRepo := thmocks.NewGroupMembersRepository()
+	groupsRepo := thmocks.NewGroupRepository(groupMembersRepo)
+	profileCache := thmocks.NewProfileCache()
+	thingCache := thmocks.NewThingCache()
+	groupCache := thmocks.NewGroupCache()
+	idProvider := uuid.NewMock()
+
+	return things.New(auth, nil, thingsRepo, profilesRepo, groupsRepo, groupMembersRepo, profileCache, thingCache, groupCache, idProvider)
+}
+
+func newServer(svc things.Service) *httptest.Server {
+	logger := logger.NewMock()
+	mux := httpapi.MakeHandler(svc, mocktracer.New(), logger)
+	return httptest.NewServer(mux)
+}
+
+func toJSON(data interface{}) string {
+	jsonData, _ := json.Marshal(data)
+	return string(jsonData)
+}
+
+func TestRemoveGroups(t *testing.T) {
+	svc := newService()
+	ts := newServer(svc)
+	defer ts.Close()
+
+	var groupIDs []string
+	for i := uint64(0); i < 10; i++ {
+		num := strconv.FormatUint(i, 10)
+		group := things.Group{
+			OrgID:       orgID,
+			Name:        "test-group-" + num,
+			Description: "test group desc",
+		}
+		grs, err := svc.CreateGroups(context.Background(), token, group)
+		require.Nil(t, err, fmt.Sprintf("unexpected error: %s", err))
+		gr := grs[0]
+
+		groupIDs = append(groupIDs, gr.ID)
+	}
+	cases := []struct {
+		desc        string
+		data        []string
+		auth        string
+		contentType string
+		status      int
+	}{
+		{
+			desc:        "remove existing groups",
+			data:        groupIDs[:5],
+			auth:        token,
+			contentType: contentType,
+			status:      http.StatusNoContent,
+		},
+		{
+			desc:        "remove non-existent groups",
+			data:        []string{wrongValue},
+			auth:        token,
+			contentType: contentType,
+			status:      http.StatusNotFound,
+		},
+		{
+			desc:        "remove groups with invalid token",
+			data:        groupIDs[len(groupIDs)-5:],
+			auth:        wrongValue,
+			contentType: contentType,
+			status:      http.StatusUnauthorized,
+		},
+		{
+			desc:        "remove groups without group ids",
+			data:        []string{},
+			auth:        token,
+			contentType: contentType,
+			status:      http.StatusBadRequest,
+		},
+		{
+			desc:        "remove groups with empty group ids",
+			data:        []string{emptyValue},
+			auth:        token,
+			contentType: contentType,
+			status:      http.StatusBadRequest,
+		},
+		{
+			desc:        "remove groups with empty token",
+			data:        groupIDs,
+			auth:        emptyValue,
+			contentType: contentType,
+			status:      http.StatusUnauthorized,
+		},
+		{
+			desc:        "remove groups with invalid content type",
+			data:        groupIDs,
+			auth:        token,
+			contentType: wrongValue,
+			status:      http.StatusUnsupportedMediaType,
+		},
+	}
+	for _, tc := range cases {
+		data := struct {
+			GroupIDs []string `json:"group_ids"`
+		}{
+			tc.data,
+		}
+		body := toJSON(data)
+		req := testRequest{
+			client:      ts.Client(),
+			method:      http.MethodPatch,
+			url:         fmt.Sprintf("%s/groups", ts.URL),
+			token:       tc.auth,
+			contentType: tc.contentType,
+			body:        strings.NewReader(body),
+		}
+		res, err := req.make()
+		assert.Nil(t, err, fmt.Sprintf("%s: unexpected error %s", tc.desc, err))
+		assert.Equal(t, tc.status, res.StatusCode, fmt.Sprintf("%s: expected status code %d got %d", tc.desc, tc.status, res.StatusCode))
+	}
+}
