+++ things/api/http/members/transport.go
@@ -1,0 +1,149 @@
+// Copyright (c) Mainflux
+// SPDX-License-Identifier: Apache-2.0
+
+package members
+
+import (
+	"context"
+	"encoding/json"
+	"net/http"
+	"strings"
+
+	log "github.com/MainfluxLabs/mainflux/logger"
+	"github.com/MainfluxLabs/mainflux/pkg/apiutil"
+	"github.com/MainfluxLabs/mainflux/pkg/errors"
+	"github.com/MainfluxLabs/mainflux/pkg/uuid"
+	"github.com/MainfluxLabs/mainflux/things"
+	kitot "github.com/go-kit/kit/tracing/opentracing"
+	kithttp "github.com/go-kit/kit/transport/http"
+	"github.com/go-zoo/bone"
+	"github.com/opentracing/opentracing-go"
+)
+
+// MakeHandler returns a HTTP handler for API endpoints.
+func MakeHandler(svc things.Service, mux *bone.Mux, tracer opentracing.Tracer, logger log.Logger) *bone.Mux {
+	opts := []kithttp.ServerOption{
+		kithttp.ServerErrorEncoder(apiutil.LoggingErrorEncoder(logger, encodeError)),
+	}
+
+	mux.Post("/groups/:id/members", kithttp.NewServer(
+		kitot.TraceServer(tracer, "create_group_members")(createGroupMembersEndpoint(svc)),
+		decodeGroupMembers,
+		encodeResponse,
+		opts...,
+	))
+
+	mux.Get("/groups/:id/members", kithttp.NewServer(
+		kitot.TraceServer(tracer, "list_group_members")(listGroupMembersEndpoint(svc)),
+		decodeListByGroup,
+		encodeResponse,
+		opts...,
+	))
+
+	mux.Put("/groups/:id/members", kithttp.NewServer(
+		kitot.TraceServer(tracer, "update_group_members")(updateGroupMembersEndpoint(svc)),
+		decodeGroupMembers,
+		encodeResponse,
+		opts...,
+	))
+
+	mux.Patch("/groups/:id/members", kithttp.NewServer(
+		kitot.TraceServer(tracer, "remove_group_members")(removeGroupMembersEndpoint(svc)),
+		decodeRemoveGroupMembers,
+		encodeResponse,
+		opts...,
+	))
+
+	return mux
+}
+
+func decodeListByGroup(_ context.Context, r *http.Request) (interface{}, error) {
+	pm, err := apiutil.BuildPageMetadata(r)
+	if err != nil {
+		return nil, err
+	}
+
+	req := listByGroupReq{
+		token:        apiutil.ExtractBearerToken(r),
+		id:           bone.GetValue(r, apiutil.IDKey),
+		pageMetadata: pm,
+	}
+
+	return req, nil
+}
+
+func decodeGroupMembers(_ context.Context, r *http.Request) (interface{}, error) {
+	if !strings.Contains(r.Header.Get("Content-Type"), apiutil.ContentTypeJSON) {
+		return nil, apiutil.ErrUnsupportedContentType
+	}
+
+	req := groupMembersReq{
+		token:   apiutil.ExtractBearerToken(r),
+		groupID: bone.GetValue(r, apiutil.IDKey),
+	}
+
+	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
+		return nil, errors.Wrap(apiutil.ErrMalformedEntity, err)
+	}
+
+	return req, nil
+}
+
+func decodeRemoveGroupMembers(_ context.Context, r *http.Request) (interface{}, error) {
+	if !strings.Contains(r.Header.Get("Content-Type"), apiutil.ContentTypeJSON) {
+		return nil, apiutil.ErrUnsupportedContentType
+	}
+
+	req := removeGroupMembersReq{
+		token:   apiutil.ExtractBearerToken(r),
+		groupID: bone.GetValue(r, apiutil.IDKey),
+	}
+
+	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
+		return nil, errors.Wrap(apiutil.ErrMalformedEntity, err)
+	}
+
+	return req, nil
+}
+
+func encodeResponse(_ context.Context, w http.ResponseWriter, response interface{}) error {
+	w.Header().Set("Content-Type", apiutil.ContentTypeJSON)
+
+	if ar, ok := response.(apiutil.Response); ok {
+		for k, v := range ar.Headers() {
+			w.Header().Set(k, v)
+		}
+
+		w.WriteHeader(ar.Code())
+
+		if ar.Empty() {
+			return nil
+		}
+	}
+
+	return json.NewEncoder(w).Encode(response)
+}
+
+func encodeError(_ context.Context, err error, w http.ResponseWriter) {
+	switch {
+	case err == apiutil.ErrBearerToken:
+		w.WriteHeader(http.StatusUnauthorized)
+	case errors.Contains(err, apiutil.ErrUnsupportedContentType):
+		w.WriteHeader(http.StatusUnsupportedMediaType)
+	case errors.Contains(err, apiutil.ErrInvalidQueryParams),
+		errors.Contains(err, apiutil.ErrMalformedEntity),
+		err == apiutil.ErrEmptyList,
+		err == apiutil.ErrMissingGroupID,
+		err == apiutil.ErrMissingMemberID,
+		err == apiutil.ErrLimitSize,
+		err == apiutil.ErrOffsetSize,
+		err == apiutil.ErrInvalidRole:
+		w.WriteHeader(http.StatusBadRequest)
+	case errors.Contains(err, uuid.ErrGeneratingID):
+		w.WriteHeader(http.StatusInternalServerError)
+	default:
+		apiutil.EncodeError(err, w)
+	}
+
+	apiutil.WriteErrorResponse(err, w)
+}
