+++ things/mocks/groups.go
@@ -30,18 +30,18 @@ type groupRepositoryMock struct {
 	profileMembership map[string]string
 	// Map of group profile where group id is a key and profile ids are values.
 	profiles map[string][]string
-	roles    things.GroupMembersRepository
+	members  things.GroupMembersRepository
 }
 
 // NewGroupRepository creates in-memory user repository
-func NewGroupRepository(roles things.GroupMembersRepository) things.GroupRepository {
+func NewGroupRepository(members things.GroupMembersRepository) things.GroupRepository {
 	return &groupRepositoryMock{
 		groups:            make(map[string]things.Group),
 		thingMembership:   make(map[string]string),
 		things:            make(map[string][]string),
 		profileMembership: make(map[string]string),
 		profiles:          make(map[string][]string),
-		roles:             roles,
+		members:           members,
 	}
 }
 
@@ -113,7 +113,7 @@ func (grm *groupRepositoryMock) RetrieveIDsByOrgMember(ctx context.Context, orgI
 	defer grm.mu.Unlock()
 
 	var grIDs []string
-	ids, _ := grm.roles.RetrieveGroupIDsByMember(ctx, memberID)
+	ids, _ := grm.members.RetrieveGroupIDsByMember(ctx, memberID)
 	for _, gr := range grm.groups {
 		for _, id := range ids {
 			if gr.OrgID == orgID && gr.ID == id {
