+++ things/service.go
@@ -114,10 +114,10 @@ type Service interface {
 	// GetGroupIDByThingID returns a thing's group ID for given thing ID.
 	GetGroupIDByThingID(ctx context.Context, thingID string) (string, error)
 
-	// Backup retrieves all things, profiles, groups, and groups roles for all users. Only accessible by admin.
+	// Backup retrieves all things, profiles, groups, and groups members for all users. Only accessible by admin.
 	Backup(ctx context.Context, token string) (Backup, error)
 
-	// Restore adds things, profiles, groups, and groups roles from a backup. Only accessible by admin.
+	// Restore adds things, profiles, groups, and groups members from a backup. Only accessible by admin.
 	Restore(ctx context.Context, token string, backup Backup) error
 
 	Groups
